requirements.txt
pyTelegramBotAPI
requests
python-dotenv
flask
psutil